chore Moving polyfills into a separate file with standard syntax to make reporting on these easier in future
