+++ elements/pfe-modal/src/pfe-modal.js
@@ -1,11 +1,7 @@
-import PFElement from "../../pfelement/dist/pfelement.js";
+// Import polyfills: startsWith
+import "./polyfills--pfe-modal.js";
 
-// StartsWith polyfill
-if (!String.prototype.startsWith) {
-  String.prototype.startsWith = function(searchString, position){
-    return this.substr(position || 0, searchString.length) === searchString;
-};
-}
+import PFElement from "../../pfelement/dist/pfelement.js";
 
 class PfeModal extends PFElement {
   static get tag() {
