+++ elements/pfe-navigation/src/polyfills--pfe-navigation.js
@@ -1,6 +1,5 @@
-// -- @TODO Needing to manually add polyfills at the moment; check into why babel is not doing this
-
 // -- POLYFILL: Array.prototype.filter
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/filter
 if (!Array.prototype.filter){
     Array.prototype.filter = function(func, thisArg) {
       'use strict';
@@ -38,94 +37,98 @@ if (!Array.prototype.filter){
       res.length = c; // shrink down array to proper size
       return res;
     };
-  }
+}
   
-  // -- POLYFILL: Element.prototype.matches
-  if (!Element.prototype.matches) {
-    Element.prototype.matches = Element.prototype.msMatchesSelector || 
-                                Element.prototype.webkitMatchesSelector;
-  }
+// -- POLYFILL: Element.prototype.matches
+// https://developer.mozilla.org/en-US/docs/Web/API/Element/matches
+if (!Element.prototype.matches) {
+  Element.prototype.matches = Element.prototype.msMatchesSelector || 
+                              Element.prototype.webkitMatchesSelector;
+}
   
-  // -- POLYFILL: Element.prototype.closest
-  if (!Element.prototype.closest) {
-    Element.prototype.closest = function(s) {
-      var el = this;
+// -- POLYFILL: Element.prototype.closest
+// https://developer.mozilla.org/en-US/docs/Web/API/Element/closest
+if (!Element.prototype.closest) {
+  Element.prototype.closest = function(s) {
+    var el = this;
+
+    do {
+      if (el.matches(s)) return el;
+      el = el.parentElement || el.parentNode;
+    } while (el !== null && el.nodeType === 1);
+    return null;
+  };
+}
   
-      do {
-        if (el.matches(s)) return el;
-        el = el.parentElement || el.parentNode;
-      } while (el !== null && el.nodeType === 1);
-      return null;
-    };
-  }
-  
-  // -- POLYFILL: Array.prototype.includes
-  if (!Array.prototype.includes) {
-    Object.defineProperty(Array.prototype, 'includes', {
-      value: function (searchElement, fromIndex) {
-  
-        // 1. Let O be ? ToObject(this value).
-        if (this == null) {
-          throw new TypeError('"this" is null or not defined');
-        }
-  
-        var o = Object(this);
-  
-        // 2. Let len be ? ToLength(? Get(O, "length")).
-        var len = o.length >>> 0;
-  
-        // 3. If len is 0, return false.
-        if (len === 0) {
-          return false;
-        }
-  
-        // 4. Let n be ? ToInteger(fromIndex).
-        //    (If fromIndex is undefined, this step produces the value 0.)
-        var n = fromIndex | 0;
-  
-        // 5. If n ≥ 0, then
-        //  a. Let k be n.
-        // 6. Else n < 0,
-        //  a. Let k be len + n.
-        //  b. If k < 0, let k be 0.
-        var k = Math.max(n >= 0 ? n : len - Math.abs(n), 0);
-  
-        function sameValueZero(x, y) {
-          return x === y || (typeof x === 'number' && typeof y === 'number' && isNaN(x) && isNaN(y));
-        }
-  
-        // 7. Repeat, while k < len
-        while (k < len) {
-          // a. Let elementK be the result of ? Get(O, ! ToString(k)).
-          // b. If SameValueZero(searchElement, elementK) is true, return true.
-          // c. Increase k by 1.
-          if (sameValueZero(o[k], searchElement)) {
-            return true;
-          }
-          k++;
-        }
-  
-        // 8. Return false
+// -- POLYFILL: Array.prototype.includes
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/includes
+if (!Array.prototype.includes) {
+  Object.defineProperty(Array.prototype, 'includes', {
+    value: function (searchElement, fromIndex) {
+
+      // 1. Let O be ? ToObject(this value).
+      if (this == null) {
+        throw new TypeError('"this" is null or not defined');
+      }
+
+      var o = Object(this);
+
+      // 2. Let len be ? ToLength(? Get(O, "length")).
+      var len = o.length >>> 0;
+
+      // 3. If len is 0, return false.
+      if (len === 0) {
         return false;
       }
-    });
-  }
-  
-  // -- POLYFILL: Event.prototype.path
-  if (!("path" in Event.prototype)) {
-      Object.defineProperty(Event.prototype, "path", {
-        get: function() {
-          var path = [];
-          var currentElem = this.target;
-          while (currentElem) {
-            path.push(currentElem);
-            currentElem = currentElem.parentElement;
-          }
-          if (path.indexOf(window) === -1 && path.indexOf(document) === -1)
-            path.push(document);
-          if (path.indexOf(window) === -1)
-            path.push(window);
-          return path;
+
+      // 4. Let n be ? ToInteger(fromIndex).
+      //    (If fromIndex is undefined, this step produces the value 0.)
+      var n = fromIndex | 0;
+
+      // 5. If n ≥ 0, then
+      //  a. Let k be n.
+      // 6. Else n < 0,
+      //  a. Let k be len + n.
+      //  b. If k < 0, let k be 0.
+      var k = Math.max(n >= 0 ? n : len - Math.abs(n), 0);
+
+      function sameValueZero(x, y) {
+        return x === y || (typeof x === 'number' && typeof y === 'number' && isNaN(x) && isNaN(y));
+      }
+
+      // 7. Repeat, while k < len
+      while (k < len) {
+        // a. Let elementK be the result of ? Get(O, ! ToString(k)).
+        // b. If SameValueZero(searchElement, elementK) is true, return true.
+        // c. Increase k by 1.
+        if (sameValueZero(o[k], searchElement)) {
+          return true;
         }
-      });
-  }
+        k++;
+      }
+
+      // 8. Return false
+      return false;
+    }
+  });
+}
+
+// -- POLYFILL: Event.prototype.path
+// https://stackoverflow.com/questions/36845515/mouseevent-path-equivalent-in-firefox-safari
+if (!("path" in Event.prototype)) {
+    Object.defineProperty(Event.prototype, "path", {
+      get: function() {
+        var path = [];
+        var currentElem = this.target;
+        while (currentElem) {
+          path.push(currentElem);
+          currentElem = currentElem.parentElement;
+        }
+        if (path.indexOf(window) === -1 && path.indexOf(document) === -1)
+          path.push(document);
+        if (path.indexOf(window) === -1)
+          path.push(window);
+        return path;
+      }
+    });
+}
