+++ elements/pfe-number/src/pfe-number.js
@@ -1,3 +1,6 @@
+// Import polyfills: isNaN
+import "./polyfills--pfe-number.js";
+
 import PFElement from "../../pfelement/dist/pfelement.js";
 import numeral from "numeral";
 
@@ -14,10 +17,6 @@ const types = {
 // use thin spaces to separate thousands chunks
 // debugger;
 numeral.locales.en.delimiters.thousands = " ";
-
-// @IE11
-// non-mutating polyfill from https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isNaN#Polyfill
-const isNaN = Number.isNaN || (n => n !== null && (n != n || +n != n));
 
 class PfeNumber extends PFElement {
   static get tag() {
